refactor: reorganize src into adapters/services/shared layers
