+++ src/adapters/editor.js
