+++ src/adapters/markdown.js
