+++ src/shared/event-bus.js
