+++ src/shared/url.js
@@ -1,0 +1,3 @@
+export const isURL = (str) => {
+  return /(?:^\w+:|^)\/\/(?:[^\s\.]+\.\S{2}|localhost[\:?\d]*)/.test(str);
+};
